music: extracted shared caching logic and test fixtures
