+++ swift/music/lib-tca/Tests/LibTCATests/AlbumDetailFeatureTests.swift
@@ -107,11 +107,3 @@ struct AlbumDetailFeatureTests {
     await store.send(.playTapped)
   }
 }
-
-private func playbackItems(album: ApprovedAlbum) -> [PlaybackItem] {
-  album.tracks.map { PlaybackItem(
-    track: $0,
-    artworkURL: album.artworkURL,
-    albumID: album.id,
-  ) }
-}
